nodes/text2speech/pyproject.toml
# SPDX-FileCopyrightText: 2022-present khulnasoft GmbH <info@khulnasoft.com>
#
# SPDX-License-Identifier: Apache-2.0
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "farm-deepstack-text2speech"
description = 'Deepstack node to convert text entities (documents, answers, etc...) into audio files.'
readme = "README.md"
requires-python = ">=3.7"
license = "Apache-2.0"
keywords = []
authors = [{ name = "khulnasoft GmbH", email = "info@khulnasoft.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "farm-deepstack",
  "numpy",
  "torch",
  "pydantic",
  "soundfile",
  "pyworld>=0.3.1",
  "espnet",
  "typeguard==2.13.3",
  "espnet-model-zoo",
  "pydub",
  "ffmpeg-python"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["pytest", "pylint", "black"]

[project.urls]
Documentation = "https://github.com/khulnasoft/deepstack-core-integrations/tree/main/nodes/text2speech#readme"
Issues = "https://github.com/khulnasoft/deepstack-core-integrations/issues"
Source = "https://github.com/khulnasoft/deepstack-core-integrations/tree/main/nodes/text2speech"

[tool.hatch.version]
path = "text2speech/__about__.py"

[tool.hatch.envs.default]
dependencies = ["pytest", "pytest-cov"]

[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=text2speech --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = ["text2speech/__about__.py"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = ["missing-module-docstring", "fixme", "R0913", "W0221"]

[tool.pylint.'DESIGN']
max-args = 7

[tool.pylint.'SIMILARITIES']
min-similarity-lines = 10
ignore-comments = true

[tool.pylint.'TYPECHECK']
# List of members which are set dynamically and missed by Pylint inference
# system, and so shouldn't trigger E1101 when accessed.
generated-members = "torch.*"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers"
markers = ["integration: integration tests", "unit: unit tests"]
log_cli = true

[tool.black]
line-length = 120
